fix auth flow and improve admin panel

+++ backend/app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean
+from sqlalchemy import Column, Integer, BigInteger, String, Boolean
 from sqlalchemy.orm import relationship
 
 from app.database import Base
@@ -8,7 +8,7 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, nullable=False, index=True)
+    telegram_id = Column(BigInteger, unique=True, nullable=False, index=True)
     full_name = Column(String(255), nullable=False)
     language = Column(String(10), default="uz")
     is_admin = Column(Boolean, default=False)
